requirements.txt
asgiref==3.7.2
Django==4.2.4
django-environ==0.10.0
djangorestframework==3.14.0
Markdown==3.4.4
psycopg2==2.9.7
psycopg2-binary==2.9.7
pytz==2023.3
sqlparse==0.4.4
